change api to MVC structure
