+++ api/routes/index.routes.js
@@ -1,0 +1,7 @@
+const usersRouter = require('./users.routes');
+
+function route(app) {
+    app.use('/users', usersRouter)
+}
+
+module.exports = route;
